refactor(web): split connector setup panels
